requirements.txt
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchreid==1.4.0
ultralytics==8.3.173
opencv-python==4.12.0.88
numpy==2.1.2
scikit-learn==1.7.1
Cython==3.1.2
gdown==5.2.0
yacs==0.1.8
pillow==11.0.0



